untitled

+++ EscolaTCC/Models/BLL/FuncCargo.cs
@@ -15,7 +15,7 @@ namespace EscolaTCC.Models
         public String dsFormacao { get; set; }
 
         [Display(Name = "Salário")]
-        public decimal No_Salario { get; set; }
+        public Int64 No_Salario { get; set; }
 
         [Display(Name = "Descrição")]
         public String Ds_Cargo { get; set; }
